# More optimize dlk runtime