--- a/dlk/python/dlk/templates/Makefile.tpl
+++ b/dlk/python/dlk/templates/Makefile.tpl
@@ -38,28 +38,21 @@ LIB_ARM_SRC := $(wildcard $(SRC_DIR)/*.S) \
     $(SRC_DIR)/func/arm_neon/batch_normalization.cpp \
     $(SRC_DIR)/func/impl/arm_neon/quantized_conv2d_tiling.cpp \
     $(SRC_DIR)/func/impl/generic/quantized_conv2d_kn2row.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/pack_16bit.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/apply_thresholds.cpp \
-    $(SRC_DIR)/matrix/arm_neon/quantized_multiplication.cpp
+    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp
 LIB_ARM_OBJ := $(patsubst %.S, %.o, $(LIB_ARM_SRC))
 LIB_ARM_OBJ := $(patsubst %.cpp, %.o, $(LIB_ARM_OBJ))
 
 LIB_FPGA_SRC := $(wildcard $(SRC_DIR)/*.S) \
     $(SRC_DIR)/func/arm_neon/batch_normalization.cpp \
     $(SRC_DIR)/func/impl/fpga/quantized_conv2d_kn2row.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp \
-    $(SRC_DIR)/matrix/arm_neon/quantized_multiplication.cpp
+    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp
 LIB_FPGA_OBJ := $(patsubst %.S, %.o, $(LIB_FPGA_SRC))
 LIB_FPGA_OBJ := $(patsubst %.cpp, %.o, $(LIB_FPGA_OBJ))
 
 LIB_AARCH64_SRC := \
     $(SRC_DIR)/func/arm_neon/batch_normalization.cpp \
     $(SRC_DIR)/func/impl/arm_neon/quantized_conv2d_tiling.cpp \
-    $(SRC_DIR)/matrix/arm_neon/quantized_multiplication.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/apply_thresholds.cpp \
-    $(SRC_DIR)/func/impl/arm_neon/pack_16bit.cpp
+    $(SRC_DIR)/func/impl/arm_neon/pop_count.cpp
 LIB_AARCH64_OBJ := $(patsubst %.S, %.o, $(LIB_AARCH64_SRC))
 LIB_AARCH64_OBJ := $(patsubst %.cpp, %.o, $(LIB_AARCH64_OBJ))
 
@@ -75,7 +68,6 @@ LIB_X86_OBJ := $(patsubst %.cpp, %.o, $(LIB_X86_SRC))
 LIB_X86_AVX_SRC := \
     $(SRC_DIR)/func/generic/batch_normalization.cpp \
     $(SRC_DIR)/func/impl/x86_avx/quantized_conv2d_tiling.cpp \
-    $(SRC_DIR)/matrix/generic/quantized_multiplication.cpp \
     $(SRC_DIR)/func/impl/generic/pop_count.cpp \
     $(SRC_DIR)/func/impl/x86_avx/apply_thresholds.cpp \
     $(SRC_DIR)/func/impl/x86_avx/pack_16bit.cpp