--- a/dlk/python/dlk/templates/Makefile.tpl
+++ b/dlk/python/dlk/templates/Makefile.tpl
@@ -68,9 +68,7 @@ LIB_X86_OBJ := $(patsubst %.cpp, %.o, $(LIB_X86_SRC))
 LIB_X86_AVX_SRC := \
     $(SRC_DIR)/func/generic/batch_normalization.cpp \
     $(SRC_DIR)/func/impl/x86_avx/quantized_conv2d_tiling.cpp \
-    $(SRC_DIR)/func/impl/generic/pop_count.cpp \
-    $(SRC_DIR)/func/impl/x86_avx/apply_thresholds.cpp \
-    $(SRC_DIR)/func/impl/x86_avx/pack_16bit.cpp
+    $(SRC_DIR)/func/impl/generic/pop_count.cpp
 LIB_X86_AVX_OBJ := $(patsubst %.cpp, %.o, $(LIB_X86_AVX_SRC))
 
 LIB_OBJ := $(patsubst %.cpp, %.o, $(LIB_SRC))