--- a/dlk/python/dlk/templates/include/operators.h
+++ b/dlk/python/dlk/templates/include/operators.h
@@ -57,6 +57,7 @@ struct binary_convolution_parameters {
   unsigned long device_input_phys_addr;
   unsigned long device_output_phys_addr;
   unsigned long device_kernel_phys_addr;
+  unsigned long device_thresholds_phys_addr;
 
   DMA_Buffer *dma_input_buffer;
   DMA_Buffer *dma_output_buffer;