<!--- 👉👉👉 Provide a general summary of your changes. 👈👈👈 -->

| |DE10-Nano (CPU)|DE10-Nano (FPGA)|Raspberry Pi 3B|
|----|----|----|----|
|before|327ms|28.7ms|528ms|
|by #353 |129ms|28.7ms|83ms|
|by #359 |94ms|19.9ms|52ms|
|by #381 |94ms|17.7ms|52ms|
|after|61.3ms|13.3ms|36.2ms|

| |Core i7-8750H@governor=performance|Xeon Gold 6154|
|----|----|----|
|before|49.8ms|100ms|
|by #365 |2.34ms@OMP_NUM_THREADS=6|7.43ms@OMP_NUM_THREADS=2|
|by #373 |1.73ms@OMP_NUM_THREADS=6|1.85ms@OMP_NUM_THREADS=18|
|after|1.42ms@OMP_NUM_THREADS=6|1.36ms@OMP_NUM_THREADS=18|

Execution time varies, these are minimum time.

## Motivation and Context
<!--- 💭💡💭 Why is this change required? What problem does it solve? 💭💡💭 -->
<!--- 🔗 If it fixes an open issue, please link to the issue here. 🔗 -->

This PR implements a lot of generic or target specific optimizations. 

## Description
<!--- 📝🎯📝 Describe your changes in detail. 📝🎯📝 -->

Optimize these elements:
- float matrix multiplication
- quantizer
- pack_input
- lookup
- extract_image_patches
- convert_tensor
- memcpy
- copy thresholds
- conv2d+apply_thresholds+pack_16bit+convert_tensor

## How has this been tested?
<!--- 🙆👌🙆 Please describe in detail how you tested your changes. 🙆👌🙆 -->
<!--- Include details of your testing environment, details of your manual tests, snippet code or commands of your manual tests, table of experiments results metrics, tests ran to see how your change affects other areas of the code, etc. -->
<!--- If you have experiments report documents please link to here. -->

```shell
$ PYTHONPATH=python/dlk python python/dlk/scripts/generate_project.py -i minimal_graph_with_shape.pb -o tmp/ -p object_detection -hq -ts -cache
$ cd tmp/object_detection.prj
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_x86_avx
$ make clean
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_fpga
$ make clean
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_arm
$ make clean
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_aarch64
$ make clean
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_x86
```

and test each executable file.

## Screenshots (if appropriate):

## Types of changes
<!--- 🏁🎌🏁 What types of changes does your code introduce? Put an `x` in all the boxes that apply: 🏁🎌🏁 -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
<!--- ✅✅✅ Go over all the following points, and put an `x` in all the boxes that apply. ✅✅✅ -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

<!--- TODO(wakisaka): After decide our code style, add this.
- [ ] My code follows the code style of this project.
-->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.