@tkng 2019-08-20
Some tests doesn't passed, something wrong happens, but I cannot understand how to execute unit test on my local pc... Please give me 3 days to understand.