@primenumber 2019-08-20
I found large array allocated on stack area, it caused stack overflow.
I fixed it.