@tkng 2019-09-10
- [x] Readabilty approval
- [x] Ownership approval

let's merge after unit tests are passed